+++ src/ragas/llms/__init__.py
@@ -1,4 +1,12 @@
-from ragas.llms.base import BaseRagasLLM, LangchainLLM, llm_factory
-from ragas.llms.llamaindex import LlamaIndexLLM
+import os
 
-__all__ = ["BaseRagasLLM", "LangchainLLM", "LlamaIndexLLM", "llm_factory"]
+from ragas.llms.base import BaseRagasLLM
+from ragas.llms.langchain import LangchainLLM
+from ragas.llms.llamaindex import LlamaIndexLLM
+from ragas.llms.openai import OpenAI
+
+__all__ = ["BaseRagasLLM", "LangchainLLM", "LlamaIndexLLM", "llm_factory", "OpenAI"]
+
+
+def llm_factory() -> BaseRagasLLM:
+    return OpenAI()
